Split into separate model and view
