+++ js/textInputView.js
@@ -1,0 +1,109 @@
+define([
+    'core/js/views/questionView'
+], function(QuestionView) {
+
+    var TextInputView = QuestionView.extend({
+
+ 		events: {
+            "focus input":"clearValidationError",
+            "change input":"onInputChanged"
+        },
+
+        resetQuestionOnRevisit: function() {
+            this.setAllItemsEnabled(false);
+            this.resetQuestion();
+        },
+
+        setupQuestion: function() {
+            this.model.setupRandomisation();
+        },
+
+        disableQuestion: function() {
+            this.setAllItemsEnabled(false);
+        },
+
+        enableQuestion: function() {
+            this.setAllItemsEnabled(true);
+        },
+
+        setAllItemsEnabled: function(isEnabled) {
+            _.each(this.model.get('_items'), function(item, index) {
+                var $itemInput = this.$('input').eq(index);
+
+                if (isEnabled) {
+                    $itemInput.prop('disabled', false);
+                } else {
+                    $itemInput.prop('disabled', true);
+                }
+            }, this);
+        },
+
+        onQuestionRendered: function() {
+            this.setReadyStatus();
+        },
+
+        clearValidationError: function() {
+            this.$(".textinput-item-textbox").removeClass("textinput-validation-error");
+        },
+
+        // Blank method for question to fill out when the question cannot be submitted
+        onCannotSubmit: function() {
+            this.showValidationError();
+        },
+
+        showValidationError: function() {
+            this.$(".textinput-item-textbox").addClass("textinput-validation-error");
+        },
+
+        // This is important and should give the user feedback on how they answered the question
+        // Normally done through ticks and crosses by adding classes
+        showMarking: function() {
+            if (!this.model.get('_canShowMarking')) return;
+
+            _.each(this.model.get('_items'), function(item, i) {
+                var $item = this.$('.textinput-item').eq(i);
+                $item.removeClass('correct incorrect').addClass(item._isCorrect ? 'correct' : 'incorrect');
+            }, this);
+        },
+
+        // Used by the question view to reset the look and feel of the component.
+        resetQuestion: function() {
+            this.$('.textinput-item-textbox').prop('disabled', !this.model.get('_isEnabled')).val('');
+
+            this.model.set({
+                _isAtLeastOneCorrectSelection: false,
+                _isCorrect: undefined
+            });
+        },
+
+        showCorrectAnswer: function() {
+            
+            if(this.model.get('_answers'))  {
+                
+                var correctAnswers = this.model.get('_answers');
+                _.each(this.model.get('_items'), function(item, index) {
+                    this.$(".textinput-item-textbox").eq(index).val(correctAnswers[index][0]);
+                }, this);
+                
+            } else {
+                _.each(this.model.get('_items'), function(item, index) {
+                    this.$(".textinput-item-textbox").eq(index).val(item._answers[0]);
+                }, this);
+            }
+            
+        },
+
+        hideCorrectAnswer: function() {
+            _.each(this.model.get('_items'), function(item, index) {
+                this.$(".textinput-item-textbox").eq(index).val(item.userAnswer);
+            }, this);
+        },
+
+        onInputChanged:function(e) {
+        	var $input = $(e.target);
+        	this.model.setItemUserAnswer($input.index(), $input.val());
+        }
+    });
+
+    return TextInputView;
+});
